Python/TileScanClient/requirements.txt
femtoapi
pyimagej
pillow